Issue #1405: Removed duplicate tests, DataStorageItem tests moved to corresponding class
